+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # @oslojs/webauthn
 
+## 0.4.0
+
+- [Breaking] Remove `verifyECDSASignature()`, `verifyRSASSAPKCS1v1_5Signature()`, `verifyRSASSAPSSSignature()`
+- [Breaking] Remove `Hash`
+- [Breaking] Remove `EllipticCurve`
+- [Breaking] Remove `ECDSAPublicKey`, `RSAPublicKey`
+- [Breaking] Remove `DecodedCOSEPublicKey`
+- [Breaking] Rename `createSignatureMessage()` to `createAssertionSignatureMessage()`
+- [Breaking] Remove `COSEPublicKey.ecdsa()`
+- [Breaking] Update `COSEPublicKey.rsa()`
+- [Breaking] Update `COSEPublicKey.algorithm()`
+- Add `COSEEllipticCurve`, `COSEKeyType`
+- Add `COSEEC2PublicKey`, `COSEOKPPublicKey`
+- Add `COSEPublicKey.isAlgorithmDefined()`, `COSEPublicKey.keyType()`, `COSEPublicKey.ec2()`, `COSEPublicKey.okp()`, `COSEPublicKey.symmetric()`
+- Add `COSEParseError`
+- No longer requires the Web Crypto API
+
 ## 0.3.0
 
 - [Breaking] Rename `AttestationFormat.Apple` to `AttestationFormat.AppleAnonymous`